Restructured using Context API.
