+++ src/App.js
@@ -5,124 +5,22 @@ import EditPost from './EditPost';
 import PostPage from './PostPage';
 import About from './About';
 import Missing from './Missing';
-import { Route, Routes, useNavigate } from 'react-router-dom';
-import { useState, useEffect } from 'react';
-import { format } from 'date-fns';
-import api from './api/posts';
-import useWindowSize from './hooks/useWindowSize';
-import useAxiosFetch from './hooks/useAxiosFetch';
+import { Route, Routes } from 'react-router-dom';
 
 function App() {
-  const [posts, setPosts] = useState([]);
-  const [search, setSearch] = useState('');
-  const [searchResults, setSearchResults] = useState([]);
-  const [postTitle, setPostTitle] = useState('');
-  const [postBody, setPostBody] = useState('');
-  const [editTitle, setEditTitle] = useState('');
-  const [editBody, setEditBody] = useState('');
-  const navigate = useNavigate();
-  const { width } = useWindowSize();
-  const { data, fetchError, isLoading } = useAxiosFetch('http://localhost:3500/posts');
-
-  useEffect(() => {
-    setPosts(data);
-  }, [data])
-  
-  useEffect(() => {
-    const filteredResults = posts.filter(post => (
-      (post.body).toLowerCase().includes(search.toLowerCase())) || 
-      (post.title).toLowerCase().includes(search.toLowerCase()))
-
-      setSearchResults(filteredResults.reverse());
-  }, [posts, search])
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    const id = posts.length ? posts[posts.length - 1].id + 1 : 1;
-    const datetime = format(new Date(), 'MMMM dd, yyyy pp');
-    const newPost = { id, title: postTitle, datetime, body: postBody}
-    try {
-      const response = await api.post('/posts', newPost);
-      const allPosts = [...posts, response.data];
-      setPosts(allPosts);
-      setPostTitle('');
-      setPostBody('');
-      navigate('/');
-    } catch(err) {
-      console.log(`Error: ${err.message}`);
-    }
-  }
-
-  const handleEdit = async (id) => {
-    const datetime = format(new Date(), 'MMMM dd, yyyy pp');
-    const updatedPost = { id, title: editTitle, datetime, body: editBody}
-    try {
-      const response = await api.put(`/posts/${id}`, updatedPost);
-      setPosts(posts.map(post => post.id === id ? { ...response.data } : post ))
-      setEditTitle('');
-      setEditBody('');
-      navigate('/')
-    } catch(err) {
-      console.log(`Error: ${err.message}`);
-    }
-    
-  }
-
-  const handleDelete = async (id) => {
-    try {
-      await api.delete(`/posts/${id}`);
-      const postsLists = posts.filter(post => post.id !== id);
-      setPosts(postsLists);
-      navigate('/');
-    } catch(err) {
-      console.log(`Error: ${err.message}`);
-    }
-  }
-
   return (
-      <Routes>
-        {/* Fixed layout (contains Header, Nav, and Footer) */}
-        <Route path="/" element={<Layout 
-            search={search} 
-            setSearch={setSearch}
-            width={width}
-        />}>
-          {/* Homepage */}
-          <Route index element={<Home 
-            posts={searchResults}
-            fetchError={fetchError}
-            isLoading={isLoading}
-          />} />
-          {/* Posts */}
-          <Route path="post">
-            {/* Homepage (index) of Posts */}
-            <Route index element={<NewPost 
-              handleSubmit={handleSubmit}
-              postTitle={postTitle}
-              postBody={postBody}
-              setPostTitle={setPostTitle}
-              setPostBody={setPostBody}
-            />} />
-            {/* Post page */}
-            <Route path=":id" element={<PostPage 
-              posts={posts}
-              handleDelete={handleDelete} />} />
-          </Route>
-          {/* Edit Post page */}
-          <Route path="edit/:id" element={<EditPost 
-            posts={posts}
-            handleEdit={handleEdit}
-            editTitle={editTitle}
-            editBody={editBody}
-            setEditTitle={setEditTitle}
-            setEditBody={setEditBody}
-          />} />
-          {/* About */}
-          <Route path="about" element={<About />} />
-          {/* Missing */}
-          <Route path="*" element={<Missing />} />
+    <Routes>
+      <Route path="/" element={<Layout />}>
+        <Route index element={<Home />} />
+        <Route path="post">
+          <Route index element={<NewPost />} />
+          <Route path=":id" element={<PostPage />} />
         </Route>
-      </Routes>
+        <Route path="edit/:id" element={<EditPost />} />
+        <Route path="about" element={<About />} />
+        <Route path="*" element={<Missing />} />
+      </Route>
+    </Routes>
   );
 }
