+++ src/Layout.js
@@ -2,13 +2,16 @@ import Header from "./Header";
 import Nav from "./Nav";
 import Footer from "./Footer";
 import { Outlet } from "react-router-dom";
+import { DataProvider } from './context/DataContext';
 
-const Layout = ({ search, setSearch, width }) => {
+const Layout = () => {
   return (
     <div className="App">
-      <Header title="Blogger" width={width}/>
-      <Nav search={search} setSearch={setSearch} />
-      <Outlet />
+      <Header title="Blogger"/>
+      <Nav />
+      <DataProvider>
+        <Outlet />
+      </DataProvider>
       <Footer />
     </div>
   )
