+++ src/PostPage.js
@@ -1,8 +1,25 @@
 import { useParams, Link } from 'react-router-dom';
+import { useContext } from 'react';
+import { useNavigate } from 'react-router-dom';
+import api from './api/posts';
+import DataContext from './context/DataContext';
 
-const PostPage = ({ posts, handleDelete }) => {
+const PostPage = () => {
+  const { posts, setPosts } = useContext(DataContext);
   const { id } = useParams();
+  const navigate = useNavigate();
   const post = posts.find((post) => (post.id).toString() === id);
+
+  const handleDelete = async (id) => {
+    try {
+      await api.delete(`/posts/${id}`);
+      const postsLists = posts.filter(post => post.id !== id);
+      setPosts(postsLists);
+      navigate('/');
+    } catch(err) {
+      console.log(`Error: ${err.message}`);
+    }
+  }
 
   return (
     <main className='PostPage'>
